Hoàn thành up result

+++ CheckMeIn/app/src/main/java/com/example/nguyenhongphuc98/checkmein/Data/db/model/ParticipantAnswerDetailsDAL.java
@@ -2,42 +2,42 @@ package com.example.nguyenhongphuc98.checkmein.Data.db.model;
 
 public class ParticipantAnswerDetailsDAL implements Comparable<ParticipantAnswerDetailsDAL>{
 
-    int num_correct;
-    int time_elapsed;
-    int total_question;
+    long num_correct;
+    long time_elapsed;
+    long total_question;
     String user_name;
 
     public ParticipantAnswerDetailsDAL() {
     }
 
-    public ParticipantAnswerDetailsDAL(int num_correct, int time_elapsed, int total_question, String user_name) {
+    public ParticipantAnswerDetailsDAL(long num_correct, long time_elapsed, long total_question, String user_name) {
         this.num_correct = num_correct;
         this.time_elapsed = time_elapsed;
         this.total_question = total_question;
         this.user_name = user_name;
     }
 
-    public int getNum_correct() {
+    public long getNum_correct() {
         return num_correct;
     }
 
-    public void setNum_correct(int num_correct) {
+    public void setNum_correct(long num_correct) {
         this.num_correct = num_correct;
     }
 
-    public int getTime_elapsed() {
+    public long getTime_elapsed() {
         return time_elapsed;
     }
 
-    public void setTime_elapsed(int time_elapsed) {
+    public void setTime_elapsed(long time_elapsed) {
         this.time_elapsed = time_elapsed;
     }
 
-    public int getTotal_question() {
+    public long getTotal_question() {
         return total_question;
     }
 
-    public void setTotal_question(int total_question) {
+    public void setTotal_question(long total_question) {
         this.total_question = total_question;
     }
